@@ -1,6 +1,6 @@
-'use strict';
-
-const parse = require('./parse.cjs');
+import { parse } from './parse.mjs';
+import { load as yamlLoad } from 'js-yaml';
+import { extname } from 'node:path';
 
 const schema = {
   type: 'object',
@@ -19,7 +19,21 @@ const schema = {
   },
 };
 
-module.exports = function loader(source) {
+function tryParseSource(src, resourcePath) {
+  const ext = resourcePath ? extname(resourcePath).toLowerCase() : '';
+
+  if (ext === '.yaml' || ext === '.yml') {
+    try {
+      return yamlLoad(src);
+    } catch (error) {
+      throw new Error(`Failed to parse as YAML: ${error.message}`);
+    }
+  }
+
+  return JSON.parse(src);
+}
+
+export default function loader(source) {
   this.cacheable();
 
   const { before, after, dereference } = this.getOptions(schema) || {};
@@ -35,7 +49,7 @@ module.exports = function loader(source) {
   }
 
   try {
-    data = JSON.parse(source);
+    data = tryParseSource(source, resourcePath);
 
     if (before) {
       data = before(data);
@@ -55,4 +69,4 @@ module.exports = function loader(source) {
   } catch (error) {
     callback(error);
   }
-};
+}
